Use nested includes in PUT to api/trip
Previously, it attempted to include both Attendee and User in
Trip.findById - this caused an error, as User isn't assosciated with
Trip. But since the frontend still expects user info to be in the
json response, I've made it use nested includes: Trip.findById includes
Attendee, and Attendee includes user.